tests: merge test_pkg_resources.py and test_pkg_resources_provider.py
Merge both `pkg_resources`-related test files into a single
`test_pkg_resources.py`.
